AJCORE-280 Clean up of the Unity unit_test
Initial work to move unit test to the IDispose design that xUnit
expects for cleanup on a test failure. Most of the test failures are
due to the fact that if a test fails it may not clean up its resources
correctly.

This change should clean up the majority of errors that have been seen
while running the unit tests.

Change-Id: I226f4dc7fd8f1b66b90e95977c45908a3f5af86a
Signed-off-by: George Nash <georgen@qce.qualcomm.com>