AJCORE-786 Reverted change to Socket.cc that broke discovery on windows.
This change reverts all of the changes made for AJCORE-746 except in the
posix/Socket.cc file. The SetReuseAddrPort has be split into the original
SetReuseAddress and SerReusePort functions.  On posix these two functions
perform the same task.

The code in SetReuseAddress is identical to the change made for AJCORE-746.

This reverts most of the changes made in commit b33e6b516e725a4c0d4cd1a38002020e1f08e761.

Change-Id: Ib1fed0e0acc33faf503fcb7361a944e2c57aa999
Signed-off-by: George Nash <georgen@qce.qualcomm.com>