AJCORE-300: Check that our advertised IPv4 address is correct before sending out advertisement.
To prevent excessive polling of the network interface addresses, the
IP name service caches the values approximately every 15 seconds.  If
we switch networks and get assigned a new IP address it is possible
that we will send out an advertisement with a stale (unconnectable)
address.  This does a quick check that the cached address is still
valid and triggers a cache update if needed before sending the
advertisement.

Change-Id: Ibbd924d504916703e0d7330ee26ad730047c0098
Signed-off-by: Todd Malsbary <tmalsbar@qce.qualcomm.com>