AJCORE-170 Add object based security APIs to Objective-C bindings
There was a previous commit for this task which was supposed to
resolve this completely. It tunrs out that the enum for
InterfaceSecurityPolicy had to be defined differently for it to
work in a foolproof way. There was an additional commit that did
not get cherry picked from the local workset which would not have
stopped the developer from using this feature but would not have been
the best approach.

Change-Id: Id879e97cdb959f53a8712844ac14fdb5a807eefe
Signed-off-by: Nikhil Dabhade <ndabhade@qce.qualcomm.com>