AJCORE-815: Restore PolicyDB code
This commit restores functionality that applies configurable policy rules to
message delivery, name ownership, and endpoint connection to the bus.  This
includes a number of bug fixes to the restored code as well as a couple of
enhancements.

The actual PolicyDB code can be included in the build by setting POLICYDB=on
on the scons command line.

Change-Id: Ica51d28ce2cec417bcb19add808e612a36e9ba84
Signed-off-by: Steve Kinneberg <stevek@qce.qualcomm.com>