Created unit tests for the AboutService
Since the about service was created as a separate project with its own
library and header files I have have chosen to create  a test that is
separate from the alljoyn_core unit tests.

With this commit we only have tests that verify the
PropertyStoreProperty class and the PropertyStoreImpl class. Additional
test will be added as work continues.

Change-Id: Ibea9ca7a0a05a8d5905aa0c3a06b56d0bddfe932
Signed-off-by: George Nash <georgen@qce.qualcomm.com>