Copied build.xml.top from alljoyn_java to top build folder
The test build script was written to be used in the top level folder.
However, before the code was given to the alliance the alljoyn_java
and alljoyn_core code were in separate git projects after
downloading the project the developer would copy the build.xml.top file
from alljoyn_java and rename it to build.xml.  Now that the projects
are combined there is no longer a need to make the developer move the
file after pulling the project.

Since the primary build for most developers is now the x86_64 build I
also changed the default CPU option from x86 to x86_64.

A copy of the original build.xml.top file is still found in the
alljoyn_java folder since many automated build scripts expect the
file to be found in that location.

Change-Id: Ie7728ced9d3787b20797670fd740798f5c1e55b1
Signed-off-by: George Nash <georgen@qce.qualcomm.com>