AJCORE-830 Presence detection feature
Support for detecting whether a peer with a well known name or
unique name is present. This is achieved by providing a public
"Ping" API which returns the status of the peer.
Added two new record types called ping and ping-reply
Added support for passing reply code in a ping reply

Change-Id: I136393a58c12c080c078a59955deaa5c96991fe2
Signed-off-by: Nikhil Dabhade <ndabhade@qce.qualcomm.com>
Signed-off-by: George Nash <georgen@qce.qualcomm.com>