Workaround added to fix Doxygen builder script
Since the only thing doxygen knows about is the output directory
it assumes the dependency tree is upto date as long as the directory
exists. Added a workaround that makes one of the output files part
of the output dependencies.  Forcing SCons to generate a dependency
graph.  Thus properly detecting changes on a rebuild.

Change-Id: I8f934e41204faff468cc521c80ca5c64688fea76
Signed-off-by: George Nash <georgen@qce.qualcomm.com>