If we are unable to allocate a new BusAttachment exit
If we are unable to allocate a new BusAttachment we should bail immediately.
Having a valid BusAttachment is critical to the sample program.  If the
BusAttachment is NULL the code tries to use it in too many locations to easily
avoid using the BusAttachment pointer.  The best solution is too exit the
program instantly and print out an error indicating the failure.

Change-Id: I616b032fa9fc0c56a6bc659035fa17ea42836340
Signed-off-by: George Nash <georgen@qce.qualcomm.com>