Updated the eclipse project file so it can find jars and libs
The original project pointed to a lib folder inside its project.
Unfortanantly this lib folder is not populated by default.  I
modified the project to read the alljoyn_java.jar and the native
alljoyn_java library from the alljoyn_java project.  The means that
both alljoyn_java and AboutFeature projects must be imported into
the eclipse project but we don't need to manualy copy library files
into another location.  The means the AboutFeature project is always
using the latest copy of the alljoyn_java library.

Change-Id: Iee072e1c18d925ae63f9aa7fec3dd1062436d5e8
Signed-off-by: George Nash <georgen@qce.qualcomm.com>