ASACORE-191 Fixed test KeyStoreListenerTest failure
The KeyStoreListener test fails when run with the gtest_repeat option.
The reason for the failure is the pointer used for the in memory keystore
is not reset to NULL after freeing the memory.  This causes this pointer
to be used next time the test is run.  This error is avoided by simply setting
the pointer back to NULL after freeing the memory.

Change-Id: I8425c97cfafc7b0f0360bd16e8fed23597e5f6cd
Signed-off-by: George Nash <georgen@qce.qualcomm.com>