ASACORE-138 Update Doxygen config file to find AboutService
Doxygen failed to include the documentation for the AboutService
if a clean build was done.  This was a result of Doxygen not knowing
that the AboutService headers needed to be installed in the
distribution before Doxygen runs.

I did not want to add complicated dependency trees to the documentation
so I just change the INCLUDE list to point directly to the source folder
not the distribution folder.

Change-Id: I75dee0de6090c8164d8cc0d9fe115c9af52cff32
Signed-off-by: George Nash <georgen@qce.qualcomm.com>