AJCORE-974 Improved the AboutClientSample
The sample now calls org.alljoyn.Bus.Ping for the device that sent the
Announce signal.

If Ping is successfull create a session and call the About methods:
org.alljoyn.About.GetObjectDescription and org.alljoyn.GetAboutData
Also read the version of the About service that sent the Announce
signal.

This change makes the AboutClientSample close to par with the C++ version
of the same sample.

The only thing the C++ sample does that this sample does not do. It calls
org.alljoyn.Icon methods and the Java sample does not.

Change-Id: I7ccae52a3120f8642cfd043adb331491c12ef1b5
Signed-off-by: George Nash <georgen@qce.qualcomm.com>