ASACORE-140 Removing old test code that is no longer used
The original task was to have the scons script copy in the
required library files from alljoyn_java build when building
into this project.

It turns out this test code has not been used by anyone for
quite some time and no one knows a good reason to keep it
around.

Change-Id: Ie63372940ad30e1e8249503946e4246134f41461
Signed-off-by: George Nash <georgen@qce.qualcomm.com>