ASACORE-163 removed sample references from header files
Many of the files included references to samples that
used the member function. The list of samples was a maintenance
problem; there was no way to make sure the samples listed were
kept upto date. The samples referenced had nothing in them that
specifically helped to explain the member function.  The code
just included the member function and that was all. Many of the
samples referenced were not even C++ and really should not have
been referenced from the C++ code.

Change-Id: Ic3824ffe0af5d772b02b9e4153a0416b463c8dc9
Signed-off-by: George Nash <georgen@qce.qualcomm.com>