ASACORE-432 Error while making numTargetLanguage ObjC call back from core on interface level translator.
This bug is caused because AJNInterfaceDescription, unlike AJNBusAttachment and AJNBusObject is
a short lived object. Once it went out of scope its AJNTranslatorImpl was deleted causing
the memory access error.

In order to fix this AJNBusAttachment now owns all AJNTranslatorImpls. This is similar to the approach
taken in the java binding.

AJNInterfaceDescription is passed a reference to the AJNBusAttachment when it is created by the
AJNBusAttachment. AJNInterfaceDescription.setDescriptionTranslator passes the AJNTranslatorImpl
to the AJNBusAttachment.

AJNBusObject is not created by the AJNBusAttachment and does not necessarily have a reference to
it. As such, the AJNBusAttachment owns the AJNTranslatorImpl until it is registered with the
AJNBusAttachment. Upon registration the AJNBusAttachment takes the ownership.

AJNTranslatorImpls reference to the AJNTranslator delegate is now __strong.

Change-Id: I89b6c116c374c546bbef21c6752e3431d2627ac4
Signed-off-by: joshuah <joshuah@qce.qualcomm.com>