ASACORE-452 AboutClient sample pings using PingAsync
When Ping was used the program would deadlock once four Announce
signals were processed at the same time because we ran out concurrent
thread handlers.  By using PingAsync this issue was avoided.

This also resulted in making a few other changes. The AboutClintSessionJoiner
no longer takes a char* but takes a qcc::String this is done because the
qcc::String is erased after creating the AboutClintSessionJoiner. So the String
must be copied into the class. For the same reason the serviceName is no longer
passed by reference but is copied into the AboutClientSessionListener.

Change-Id: If78305d05e602e49ce396816e9615188c90f475e
Signed-off-by: George Nash <georgen@qce.qualcomm.com>