ASACORE-259 Calling AboutIconClient.GetContent no longer returns an invalid pointer
The problem was that a pointer to memory that was only allocated in the GetContent
member function was returned.  This pointer was no longer valid upon return. To
maintain the API design the GetContent function now allocates enough space for the
icon content and copies the icon into the memory.  This causes a problem that now
we are allocating memory and the user must free that memory.  This is a design
pattern that we have tried to avoid because it leads to memory leaks since the user
often does not know that they need to free the memory.

There is a new method GetIcon that take an AboutIconClient::Icon class that will
obtain a copy of the icon context.  When the AboutIconClient::Icon goes out of
scope the memory is freed.

As part of this change I added a call to AboutIconClient.GetIcon function to the
AboutClient sample. I also cleaned up the console output of the AboutClient
sample to make it more readable.

Change-Id: Id7fc85e9c7197b483316669fb3b54183b36bfc75
Signed-off-by: George Nash <georgen@qce.qualcomm.com>