ASACORE-429 AboutClient will attempt to Join a session even on failed Ping
Since Older buses did not implement ping we could still receive a failure response
to the ping method on services we could Join.  The samples will attempt to join a
session even if the ping failed.  The comments make it clear that if all services
are running a newer version of AllJoyn then they can skip attempting to join a
session.

Change-Id: Ica4a49d566950b2db0ba66c0e7052ccf07f28ad4
Signed-off-by: George Nash <georgen@qce.qualcomm.com>