ASACORE-538 fixed issue preventing jUnit running on gingerbread
The value returned by NewWeakRef is a jweak object that was being
passed into the GetObjectClass which expects a jobject.

Since the clazz object was only used locally I used jobj instead
of the jweak object jtranslator.

The Translator is no longer a static variable

Making the Translator a static variable that was assigned
the same time the alljoyn_java library was loaded was
causing library load errors on gingerbread.

Change-Id: I294001e413d849ca7a0b633b3e31ef19da8f0917
Signed-off-by: George Nash <georgen@qce.qualcomm.com>