ASACORE-552 Fixed multiple memory leaks from About samples
The InternalAnnounceHandler was creating a std::pair on the heap
and not freeing them.  There were a few other coding errors that
resulted memory leaks found in the samples.  even though its not
required BusAttachment.Join() is called after stop to make sure
the BusAttachment has stopped running before the application exits.
This is done so that I knew the BusAttachment had completed running
when valgrind printed its results.

Also updated the location of the ';' character for the GetContent
method.  The code thought both GetContent and GetIcon were
deprecated when only GetContent was deprecated.

Change-Id: Ie7bff987e8e3aaf11b0a3640e643817a8dcf5f55
Signed-off-by: George Nash <georgen@qce.qualcomm.com>