AJCORE-938 Fixed possible undefined behavior code path
When the code creates a C string from a jstring its possible that
the resulting C string is a NULL pointer.  When the NULL pointer
is compared to qcc::String it ultimately was being compared using
the C strcmp function call.  The behavior of strcmp is undefined
for NULL pointers.

To prevent this check the calls to JString::c_str for NULL before
doing the comparison operation.

Found using static analysis tool.

Change-Id: I5a2e8677ecaf590e6a939e5654e4bdd7954e4c03
Signed-off-by: George Nash <georgen@qce.qualcomm.com>