ASACORE-584 Put all build output in the top level build folder
The location of the top level build folder is decided by the location
of the SConstruct file that was run.  For example if you run SCons from
`alljoyn/alljoyn_core` then all of the output will endup in
`alljoyn/alljoyn_core/build`. If SCons is run from `alljoyn` then all
of the output will endup in `alljoyn/build`

Change-Id: Ib13e65d5de6e3fbe7d00e7fcd93bcd34c2ed31a4
Signed-off-by: George Nash <georgen@qce.qualcomm.com>