ASACORE-591 Aboutclient Java sample now leaves the session
The AboutClient Java sample had a coding error that meant
once the client had Join a session it never left the session
even after it was done with the session.

The code to leave the session was inside a catch block when
it should have been outside the catch block.

Change-Id: I6b3703f4d586daa8554723421f5b2157c3044f96
Signed-off-by: George Nash <georgen@qce.qualcomm.com>