ASACORE-633 Fix for race condition in java Announce signal
Need to make sure the AddMatch rule is only applied after the
interfacelist is set up.  If the AddMatch rule is done first
its possible to recieve the Announce signal before the interface
is in the interfacelist then the code will wrongly think that
no interface matches with the Announce signal recieved.

Change-Id: I5f8754ceeaa6cf4ab7e4651e4c73911761c3c3db
Signed-off-by: George Nash <georgen@qce.qualcomm.com>