ASACORE-648 Make Java methods start with a lower case
Java programming convention states that methods and variables
should start with a lower case letter.

This is such a wide spread convention that many tools including
static analysis tools look for capitalization issues with Java
methods. Using capital letters for method names is seen as a
maintance issue since the method looks like a class to programmers.

Here we have a small problem.  AllJoyn interfaces by convention
start methods and signals with a capital letter. This
convention is inherited from the DBus specification.  Luckly
we have a way to have the AllJoyn interface use a capital
letter and the Java interface have a lower case letter.  This
can be done by specifying the `name` of the interface in the
BusMethod or BusSignal annotation.

This commit changes most instances that a method was started with
a capital letter to use a lower case letter to addhere with Java
coding conventions.

Change-Id: Ib5f0d9930649d50b961948a785a76a655548ff01
Signed-off-by: George Nash <georgen@qce.qualcomm.com>