AJCORE-1208 Addressed possible syncronisation issue in unit test
The boolean varible is used to tell if the onPingListener is called
however the variable is synchronized inconsistently.  This is not
likly to cause any problems with the way the test is writen.

Although its unlikly to cause problems the fix is simple add
syncronisation into the onPinListener code.

Issue found using static analysis tool.

Change-Id: I7f31a66a71ee42087bd0d1ece82a0975da9fdac2
Signed-off-by: George Nash <georgen@qce.qualcomm.com>