ASACORE-651 Fixed possible race condition when using Announce ObjectDescriptions
Mutex guards were added to make the code thread safe however two
sections were missed in the last commit.  This adds mutex locks around
two additional blocks of code that should have mutext locks.

The locks were added for code that added or removed things from the
m_AnnounceObjectsMap but it missed the locations that the map was
iterated through to place that infomation into MsgArgs.

Change-Id: I1befcd498eadc4020068f66e33565334a1773641
Signed-off-by: George Nash <georgen@qce.qualcomm.com>