ASACORE-710 Unregistering an AnnounceHandler also calls RemoveMatch
Before this change, registering an AnnounceHandler also added an
AddMatch rule to the rules table.  Unregistering the AnnounceHandler
failed to removed that match rule from the rules table. The code
that maped the signal to the AnnounceHandler was cleared. This made
it so the AnnounceHandler was not called but the router node was
still activly trying to find any device implementing the interfaces
specified in the match rule.

Now if an AnnounceHandler is unregistered then the RemoveMatch is called
for the match rule that was added when the handler was registered.

Change-Id: Iaab7cf1bde26ce0df37cdd06e66f7246570d41b5
Signed-off-by: George Nash <georgen@qce.qualcomm.com>