ASACORE-670 The AboutService is part of core and should always build
This change removes the need to explicitly specify about in the list
of services. In addition the C++ version of the About service is built
from alljoyn_core.  The BINDINGS=cpp option is is no longer required
to build the AboutService or its samples.  The BINDINGS=java is still
required to build the Java version of the service. The java version
of the service is still built the same.

Change-Id: I3101c978d3c44911932bf90abb82709e5f50cf07
Signed-off-by: George Nash <georgen@qce.qualcomm.com>