ASACORE-750 Fixed deadlock in Remove announce handler code
If an announce signal was recieved at the same time as the
remove handler code was called the code could deadlock.  This
was a result of the handler waiting on a lock held by the
Handler removal code and the RemoveMatch call found in the
remove handler code is blocked by the announce handler holding
onto the reply port.

This change makes sure the announceMapLock is not being held
when RemoveMatch is called.

Change-Id: I9fe5617094564137bbc2564683ec614ec94cd23d
Signed-off-by: George Nash <georgen@qce.qualcomm.com>