ASACORE-758 Fixed segfault that resulted from pinging undiscovered unique name
ASACORE-742 Bus timeut for a previously discovered node is reported as
ALLJOYN_PING_REPLY_TIMEOUT rather than BUS_REPLY_IS_ERROR_MESSAGE

Changed error code ER_ALLJOYN_PING_REPLY_UNIMPLEMENTED to
ER_ALLJOYN_PING_REPLY_INCOMPATIBLE_REMOTE_ROUTING_NODE

If the name was not discovered it will not be found in the m_peerInfoMap. When
the RefreshCache method iterates through the map the value is never found so the
iterator that should hold the name remains set at m_peerInfoMap.end() when the
code tries to use the iterator it segfaults.

While fixing I saw that the documentation for Ping was inaccurate. It will never
return ER_BAD_ARG_1 if a NULL pointer is passed in.  It will return
ER_BUS_BAD_BUS_NAME so I fixed the documentation.

Change-Id: I4aa3f397b74603b18b62a74ef61aff47c8863972
Signed-off-by: George Nash <georgen@qce.qualcomm.com>
Signed-off-by: Todd Malsbary <tmalsbar@qce.qualcomm.com>