ASACORE-775 About feature rework
The AboutService was originally developed as a separate service from the core
functionality of AllJoyn. Now the AboutService is a core requirement for AllJoyn.

All new code is in the `ajn` namespace not the `ajn::services` namespace.

`AnnounceHandler` will become `AnnounceListener`

In the core code all callbacks are handled by listeners not by handlers. By calling it a
listener we hold consistent with the naming scheme used in alljoyn_core.

Functionality of the AnnouncementRegistrar was moved into BusAttachment.

`AnnouncementRegistrar::RegisterAnnounceHandler` -> `BusAttachment.RegisterAboutListener`
`AnnouncementRegistrar::UnRegisterAnnounceHandler` -> `BusAttachment.UnregisterAboutListener`
`AnnouncemenetRegistrar::UnRegisterAllAnnounceHandlers` -> `BusAttachment.UnregisterAllAboutListeners`

Functionality of the AboutService was moved into AboutObj. AllJoyn already has a
terminology for code that implements an interface. It is a BusObject not a Service.

Functionality of the AboutClient was moved into AboutProxy. Once again AllJoyn already
has terminology for the remote object that calls a BusObject it is a ProxyBusObject.

`AboutClient.GetObjectDescripitons` -> `AboutProxy.GetObjectDescriptions`
`AboutClient.GetAboutData` -> `AboutProxy.GetAboutData`
`AboutClient.GetVersion` -> `AboutProxy.GetVersion`

The ObjectDescription information will no longer be a Public member variable of the
`AnnounceHandler` as `std::map<qcc::String, std::vector<qcc::String> >` but will be
maintained as a separate class `AboutObjectDescription`. This allows us to hide the
internals of the ObjectDesctipion making it simpler to write language bindings. It also
lets us write member functions to help interact with the Object description.

Replace `AboutPropertyStore` and `AboutPropertyStoreImpl` with a class `AboutData` that
handles the same functionality without exposing access to information that is specified
in the AboutService specification.

Move the About interfaces into the `AllJoynObj`. The interfaces in question are
`org.alljoyn.About` and `org.alljoyn.Icon` since the interfaces are part of alljoyn_core
they will no longer be announced. This commit only contains the BusObject and
ProxybusObject for the `org.alljoyn.About` interface.

Change-Id: I65f21189672f038e433b299477b8c7d807fe73e5
Signed-off-by: George Nash <georgen@qce.qualcomm.com>