ASACORE-775 About feature rework unit tests and improvements
When trying to discover who is implementing an interface its expected that a
common use case is going to be interest in only a single interface.  For that
reason a version of the RegisterAboutListener/UnregisterAboutListener were
added that can accept a single interface.

Removed code from the AboutDataTest unit test for the CreateFromXml member
function test. The tests were testing something that was unimplemented yet.

Added AboutListenerTest unit test code this is a direct mirror of the
AnnounceHandlerTest code that is found in Original About Service
implementation.

Fixed wildcard interface matching in the AboutObjectDescription.HadInterface
member function. This error was found using the AnnounceHandlerTest unit test.

Change-Id: Id7fe3b5b0224b549dba406aec467c27291eb229b
Signed-off-by: George Nash <georgen@qce.qualcomm.com>