ASACORE-775 About feature rework Update AboutObj to use BusObjects that announce
when AboutObj.Announce is called it will add any BusObjects that were marked as
announced to the AboutObjectDescription that is announced. This is always done
and is not optional.

BusAttachment.GetAboutObjectDescription will always clear the previous contents
of an AboutObjectDescription passed in and replace the contents with the current
state of the registered BusObjects.

The AboutObj no longer holds references to the AboutData and AboutObjectDescription
it now holds a copy of the information.

Minor updates to existing test to account for current changes.

Change-Id: Ic60242e5889b4afc9b3d7a47dc403ca7fdce0366
Signed-off-by: George Nash <georgen@qce.qualcomm.com>