ASACORE-594 Fix multiple build warnings reported when building for Java
Most of the warnings fall in one of three categories.
 - Unused variable/method
 - Unused import
 - Raw type when Generic class was expected

The raw types (i.e. class) were changed to generic classes (class<?>)

Unused imports were removed.

Many of the methods that were reported as unused were actually called via jni
I added a suppress warning for those methods.

Many of the unused variables were in unit tests that did not care what the
return value was just that the method call did not fail.  For most of those
I added an assertTest.  Some we expected to throw an exception so I added a
fail to the unit test if the next line of code was run. The fail statement
would use the "unused" variable.

Change-Id: I5cc7f763a88371641d2ea077823c4c09c27c3d96
Signed-off-by: George Nash <georgen@qce.qualcomm.com>