ASACORE-775 About Feature rework fixed issues identified in code review
This commit contain multiple fixes for issues identified in code review.

Biggest change. The Announced member function in the AboutListener passes
in const variables. Several small changes were made to the
AboutObjectDescription to make it so its helper functions could be used
even from a const object.

Changed BusObject member function GetAnnouncedInterfaces to
GetAnnouncedInterfaceNames.

Fixed the name of the include guards to match the style used in the rest of the
alljoyn code base.

Cleaned up documentation.

Change-Id: I50b089038ef779f73d6be5b55c89296c4df9f352
Signed-off-by: George Nash <georgen@qce.qualcomm.com>