ASACORE-775 About Feature rework - AddInterface now takes AnnounceFlag
The AddInterface member function for BusObject took a raw boolean value as the
isAnnounced variable.  It now changed to an enumerated flag that gives meaning
now the isAnnounced can be set to HUSHED or ANNOUNCED.  This gives more meaning
than true or false in this member functions context.

This commit also update the documentation warnings reported when building
the API documentation.

Change-Id: I898dbed6b2d95834dadba07b0b85523bb662000c
Signed-off-by: George Nash <georgen@qce.qualcomm.com>