Implement Semaphore for posix
Note that Darwin is not considered a fully posix compliant system
It does not support sem_init for instance.
A possible improvement would be to refactor this commit using
sem_open() as this is also supported by Darwin. This is not the
only problem though. clock_gettime() is another one for instance.

Change-Id: Ia41d913b433d1d709730ee064c798cf5934459dd
Signed-off-by: Meggan Ehret <meggan@qeo.org>