ASACORE-775 About Feature rework - Add AboutDataListener
This commit contain two major changes. It contains Adds the
AboutDataListener class. This class is a pure virtual funtion that
serves as an interface that must be implemented to send out an Announce
signal.

The two virtual member functions found in the AboutDataListener are:
  GetMsgArg - This is member function is responsible for creating the
              MsgArg that is returned when a user calls
              org.alljoyn.About.GetAboutData. The returned MsgArg must
              contain the AboutData dictionary for the Language
              specified.
  GetMsgArgAnnounce - This member funtion is responsible for creating
              the MsgArg that is sent out in the
              org.alljoyn.About.Announce signal. This is always
              localized using the DefaultLanguage.

The AboutData class is now an implementation of the AboutDataListener.

When sending an Announcement the AboutDataListener is passed in as a
variable not AboutData.

The AboutListener no longer passes an AboutObjectDescription and
AboutData into the Announced callback function.  Instead it passes the
MsgArg those then can be converted to AboutObjectDescription or
AboutData by the user. The AboutProxy methods GetAboutData and
GetObjectDescription also return MsgArgs to be consistant.

Updated Samples and unit tests to account for the changed return value
for the Announced method.

The AboutData and AboutObject Initialize functions have been renamed to
CreateFromMsgArg.

Some additonal cleanup. Fixing spelling errors and removing unused code.

Change-Id: Ie81558820fd1e8712902307655d410bbaa511aff
Signed-off-by: George Nash <georgen@qce.qualcomm.com>