ASACORE-775 GetAboutObjectDescription changed to GetAnnouncedObjectDescription
The member function GetAboutObjectDescription was renamed to
GetAnnouncedObjectDescription a name more descriptive to what the function
is responsible for doing.

The member function GetAnnouncedObjectDescription returns a
MsgArg not an AboutObjectDescription the idea being that the
AboutObjectDescription class should be just a helper class.

Alljoyn methods actually return MsgArgs when a MsgArg is expected
not some other class.  This is more future proof and more consistant
with the directions the About feature rework is going.

Change-Id: Id1bde53583ac573504dc9c6684f8397b82bc346c
Signed-off-by: George Nash <georgen@qce.qualcomm.com>