ASACORE-775 About Feature rework - you must have an object to announce
The AboutObj no longer has the option to AboutObjectDescription when
making an announcement. The only way that a BusObject will be announced
is if that BusObject is registered with the BusAttachement.

Since many of the unit test used the fact that you could build a
AboutObjectDescription for Objects you did not really own I had to
update the unit tests and test programs.

Change-Id: I2765c2747b32372e74c4af71e172e38ae1ad69f6
Signed-off-by: George Nash <georgen@qce.qualcomm.com>