ASACORE-775 About Feature rework - add option to announce AboutObj
Rather than force every user to adjust there programs that expect the
org.alljoyn.About interface to be part of the announced object
descriptions we not give the user an option to add the org.alljoyn.About
interface to the announcement. This option defaults to UNANNOUNCED since
there is no value added by announcing the interface other than backwards
compatability.

Change-Id: I4c0dd2600b4c502d2c0c8982ca7467f97f912521
Signed-off-by: George Nash <georgen@qce.qualcomm.com>