ASACORE-775 About Feature rework - removed un-needed functions
Since the AboutObjectDescription is only used as a helper function its
no longer needed to build the MsgArg that is used in Announcements.
This makes many of the functions contained in the AboutObjectDescription
class useless. The Add, Remove, and Merge functions have been completely
removed or made private if they were used by the class its self.

This also adds a constructor that can create an AboutObjectDescription
class from a MsgArg.

Update the unit test to take into account the changes to the class.

Change-Id: I9f5366263f7f9207cc64a07acdcfe9a05a154cf6
Signed-off-by: George Nash <georgen@qce.qualcomm.com>