ASACORE-775 About Feature rework - automatically add localized values
Localized values are implicitly added to the SupportedLanguages. When a
localized value is added to the AboutData the language tag specified for
the localized value is automatically added to the list of supported
languages.  This is also true for the CreatUsingXml method.

In addition there is no need to specify the AjSoftwareVersion from in
the xml when creating the AboutData from xml it will automatically be
filled in.

Change-Id: Ic5ea99fd640adf5359e0d4307238eb5d7e23c9f4
Signed-off-by: George Nash <georgen@qce.qualcomm.com>