ASACORE-775 About Feature rework - split the RegisterListener
Split the RegisterAboutListener into two parts.  One that registers
the listener. Another that tells what interfaces the user is interested
in using.

This change has a few effects. It simplifies the code for handling
announce signal drastically making it less error prone. This
alse means there is no filtering of announce singnals based on
the interfaces contained in the singal other than the match rule
its self. This puts the responsibility on the developer using
the about interface but we have helper functions in the
AboutObjectDescription class to help the developer.

Change-Id: Ia9f29c47fa30f49636f3e49b9d4405978f90264b
Signed-off-by: George Nash <georgen@qce.qualcomm.com>