ASACORE-928 Fix Windows (debug) build break and App Verifier.
- FileStream.cc when built by Visual Studio 2013 compiler in debug mode causes
a build break due to DeleteFile being defined as DeleteFileA by windows.h.
The fix is to add scope operator to identify that the function is intending
to call the system API and not recursion.

- Environ.cc contains a defect when the project is built with non-UNICODE
flavor, which could lead to a run-time crash. The function assumes the call is
to the unicode ('W') version of system (environment) APIs, which is not the
case when the project does not have UNICODE defined (default). Since the
function is targeting the unicode version, the system calls need to reflect
the same – so the fix is to add 'W' suffix to these system calls.
This was caught by Microsoft App Verifier tool.

Change-Id: I33e76a461c531804c7bd61be83054514290c583d
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>