ASACORE-917 All-session broadcast signal mechanism
This introduces a new signal delivery mechanism for AllJoyn. Signals
emitted with the ALLSESSION_BROADCAST flag will be sent to all peers
that are joined in at least one session hosted by the emitting peer. The
signal is a broadcast signal, and as such subject to match rules.

On legacy routers that do not support this mechanism, the all-session
broadcast devolves to a classic global broadcast.

Change-Id: I10156c01153e87aee3e36c340ac24f4d23a9b363
Signed-off-by: Meggan Ehret <meggan@qeo.org>