ASACORE-775 About feature rework - updated AboutIconProxy
Updated the AboutIconProxy class to match the same design used in the
AboutProxy class. Passing in the remote bus name and the sessionId are
done as part of the constructor. This makes making method calls clearer
since you don't have to pass in as much information each call.

This addresses ASACORE-992

Change-Id: I73b3b0e0d7ecd84167fcba2f19667754f5a9a3e9
Signed-off-by: George Nash <georgen@qce.qualcomm.com>