ASACORE-775 About Feature Rework
Additional member functions to access the values of a field.
 - IsFieldRequired: used to find out if the field is a
                    required filed
 - IsFieldAnnounced: used to find out if the field is a part
                     of the announced field data
 - IsFieldLocalized: used to find out if the field should be
                     localized
 - GetFieldSignature: Used to get the MsgArg signature for the
                      field.

In addition to the member variable m_aboutFields was moved from the
protected member variables to private.  This fits in with AllJoyn's
API design of not exposing stl containers as part of the public API.

Moving the m_aboutFields was possible because of the added member
functions.

Created the member function InitializeFieldDetails while writing the
unit tests for the access functions I discovered that the
SoftwareVersion field had an incorrect value initialized. When I had
to change that value in three places I knew it would be less error
prone to put the initialization of the functions into a single member
function.

New test added to verify the values of each field using the newly
added accessor functions

Updated the AboutDataTestAboutData class to use the accessor
functions in place of the member variable m_aboutFields.

Change-Id: Ie27c90f51350d439d223cfe7b18a4ea7b9f11618
Signed-off-by: George Nash <georgen@qce.qualcomm.com>