ASACORE-775 About Feature rework
Changed the names of the member functions in the AboutDataListener

GetMsgArg was changed to GetAboutData
GetMsgArgAnnoune was changed to GetAnnouncedAboutData

Although the original names were good.  I feel the changed name is
more descriptive about whats actually returned.  MsgArg could be
anything while AboutData in terms of the About feature have a definite
meaning.

All of the other changes were in response to this name change.

Change-Id: Ic4df029f5f652eec149f480c371e63610e68bf4b
Signed-off-by: George Nash <georgen@qce.qualcomm.com>