ASACORE-775 About Feature rework - API cleanup
Modification to improve readability of code.

Modified the FieldDetails struct to take a mask of the fieldDetails
instead of having a boolean value for each field. When creating a new
field you can read the name of the field values. Not just a chain of
true false values.

Rather than try and read the mask every time we want to know if a
value is set now we call the IsField member functions that return the
boolean based on the masked value.

Change-Id: Ica7925044ecf0bdcf039af147a111e196c5ac3ec
Signed-off-by: George Nash <georgen@qce.qualcomm.com>