ASACORE-775 Fixed compiler error
When building the release variant we would get an unused variable
build warning because the QCC_DbgPrintf does not print in release
mode.  In the code only the the debug version was using the status
variable with this modification the variable is used in release
build as well as the debug build.

Change-Id: I7bcaa391084f826b527c3a1e2dabcc0be6110973
Signed-off-by: George Nash <georgen@qce.qualcomm.com>